backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.21.1
supabase
python-dotenv==1.0.0
